Move node-realted files into 'node' folder.
