+++ console/src/app/core/models/node/mongoose-run-node.model.ts
@@ -1,4 +1,4 @@
-import { ResourceLocatorType } from "./address-type";
+import { ResourceLocatorType } from "../address-type";
 
 export class MongooseRunNode {
